Improve response format

+++ src/main/java/it/gagagio/bondsearchtool/controller/response/BondResponse.java
@@ -3,10 +3,11 @@ package it.gagagio.bondsearchtool.controller.response;
 import it.gagagio.bondsearchtool.model.BondType;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Builder
 public record BondResponse(String isin, String name, MarketResponse market, Instant maturityAt, boolean perpetual,
-                           Integer coupon, Integer lastPrice, CountryResponse country, Integer yieldToMaturity,
+                           BigDecimal coupon, BigDecimal lastPrice, CountryResponse country, BigDecimal yieldToMaturity,
                            BondType type) {
 }
